Address possible UAF if a future is called before a callback
This adds `CallbackContext` which wraps `CallbackInner` in an `Arc`.

The pointer given to the callbacks is a clone of the Arc, thus the allocation won't be released until the callback drops the final ref.

An atomic cancelled flag is set by `Drop` in the CallbackContext which is checked by the callbacks to just drop the Arc and return.

Another advantage is all shared data is using interior mutability atomics or mutex, so overall it should be safer as the callback is likely called from a different thread.